Fix build warnings in render subsystem headers and sources (#291)
MSVC /W4 build was emitting 15 warnings across 8 distinct sites in

the render subsystem (per the pilot snapshot in

doc/build-warnings-pilot.md). This change drives the count to zero.

- meshcomponent.h: getVertexCount() / getIndexCount() narrowed

  size_t -> uint32_t implicitly. Wrap the size() call in

  static_cast<uint32_t>(...) to silence C4267 explicitly. The

  Vulkan command buffers consume these values as uint32_t already,

  so semantics are unchanged. Collapses 8 raw warning instances at

  once (the header is included by 4 TUs).

- vulkanapi.cpp: submitDrawCall, createBuffer, uploadBuffer are stub

  bodies with parameters reserved for future-phase wiring. Mark them

  [[maybe_unused]] to silence C4100 without falsely renaming the

  parameters or casting to void. Same treatment for ensureGlyph's

  pixelSize parameter in rendercomponent.cpp.

- WIN32_LEAN_AND_MEAN was defined globally via add_compile_definitions

  in the top-level CMakeLists. The bundled FreeType target re-defines

  the macro internally, producing C4005 redefinition on Windows.

  Drop the macro from the global definitions block and push it onto

  the vigine target (and example-window target) PRIVATE so the

  external/freetype subdirectory does not inherit it on its compile

  command line.

Build verified clean at /W4 on MSVC 19.50 (VS 18 Insiders) for both

Debug and Release. ctest 197/197 green. example-window.exe launches

and runs.

+++ include/vigine/ecs/render/meshcomponent.h
@@ -49,8 +49,8 @@ class MeshComponent
     const std::vector<Vertex> &getVertices() const { return _vertices; }
     const std::vector<uint32_t> &getIndices() const { return _indices; }
 
-    uint32_t getVertexCount() const { return _vertices.size(); }
-    uint32_t getIndexCount() const { return _indices.size(); }
+    uint32_t getVertexCount() const { return static_cast<uint32_t>(_vertices.size()); }
+    uint32_t getIndexCount() const { return static_cast<uint32_t>(_indices.size()); }
 
     // Procedural geometry in shader (e.g., cube.vert, sphere.vert generate vertices)
     void setProceduralInShader(bool procedural, uint32_t vertexCount = 0)
